added bookreducer
