+++ src/index.js
@@ -3,17 +3,14 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App';
 import { Provider } from 'react-redux';
+import { createStore } from 'redux'
+import { booksReducer } from './reducers/books';
 
-const initialState = [
-  { id: 0, title: "naruto", category: "action" },
-  { id: 1, title: "onepiece", category: "action" },
-  { id: 2, title: "casa", category: "biography" },
-  { id: 3, title: "doctor", category: "horor" },
-  { id: 4, title: "league", category: "history" },
-];
+
+const store = createStore(booksReducer)
 
 ReactDOM.render(
-  <Provider store={initialState}>
+  <Provider store={store}>
     <App />
   </Provider>,
   document.getElementById('root'),
